Struct ethminer::Miner

pub struct Miner {
    // some fields omitted
}

Keeps track of transactions using priority queue and holds currently mined block.

Methods

impl Miner

fn new() -> Arc<Miner>

Creates new instance of miner

fn set_author(&self, author: Address)

Set the author that we will seal blocks as.

fn set_extra_data(&self, extra_data: Bytes)

Set the extra_data that we will seal blocks with.

fn set_gas_floor_target(&self, target: U256)

Set the gas limit we wish to target when sealing a new block.

fn set_minimal_gas_price(&self, min_gas_price: U256)

Set minimal gas price of transaction to be accepted for mining.